The html assignment
I worked in repl.it to write the code and see the results. Here it is: https://repl.it/@oanafatu/SmartCodinghtml
I also shared it via hit gist here: https://gist.github.com/oanafatu
I am not a pro in Git, so I just hope you will be able to access the code and check what I've done.